requirements.txt
fastapi
uvicorn
python-multipart
jinja2
aiofiles
google-genai
pydantic
